+++ packages/parcel-transformer/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @compiled/parcel-transformer
 
+## 0.11.0
+
+### Minor Changes
+
+- 966a1080: Changed the approach of stylesheet extraction on Parcel
+
+  - @compiled/parcel-resolver is no longer used
+  - Pass styleRules to optimizer via metadata
+  - Optimizer then collects styleRules, and inserts it to output HTML
+
+### Patch Changes
+
+- Updated dependencies [966a1080]
+  - @compiled/babel-plugin-strip-runtime@0.17.0
+
 ## 0.10.0
 
 ### Minor Changes